refactor(configs): move Default impls into defaults module
