+++ daemon/configs/src/lib.rs
@@ -4,16 +4,17 @@
 
 #![warn(missing_docs)]
 
-use crate::shortcuts::{Action, Binding, Key, KeyChord, Modifiers, Prefix, Shortcuts};
-use crate::theme::Theme;
-
 pub mod error;
 pub mod shortcuts;
 pub mod theme;
+mod defaults;
 pub(crate) mod path;
 pub(crate) mod raw;
 
 pub use error::{OzmuxConfigsError, OzmuxConfigsResult};
+
+use crate::shortcuts::Shortcuts;
+use crate::theme::Theme;
 
 /// Fully-resolved ozmux configuration.
 #[derive(Clone, Debug)]
@@ -22,37 +23,4 @@ pub struct OzmuxConfigs {
     pub shortcuts: Shortcuts,
     /// Theme configuration.
     pub theme: Theme,
-}
-
-impl Default for OzmuxConfigs {
-    fn default() -> Self {
-        Self {
-            shortcuts: Shortcuts {
-                prefix: Prefix {
-                    chord: KeyChord {
-                        key: Key::Char('b'),
-                        modifiers: Modifiers {
-                            ctrl: true,
-                            ..Default::default()
-                        },
-                    },
-                    timeout_ms: 2000,
-                },
-                bindings: vec![Binding {
-                    chord: KeyChord {
-                        key: Key::Char('x'),
-                        modifiers: Modifiers::default(),
-                    },
-                    action: Action::ClosePane,
-                }],
-            },
-            theme: Theme {
-                background: "#1a1b26".into(),
-                foreground: "#c0caf5".into(),
-                accent: "#414868".into(),
-                border: "#414868".into(),
-                destructive: "#f7768e".into(),
-            },
-        }
-    }
 }
